# api-gateway/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MessageBox API Gateway",
    "description": "API gateway to interact with the MessageBox backend micro-services set.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server",
      "variables": {}
    }
  ],
...
  "paths": {
...
    "/Roles/{id}": {
      "get": {
        "tags": ["Role"],
        "summary": "Get the specific details by the role id",
        "description": "Get all the details about the role by the role id.",
        "operationId": "GetRoleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique Role ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full role details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "put": {
        "tags": ["Role"],
        "summary": "Update the role by the id",
        "description": "Update the role by the id.",
        "operationId": "UpdateRoleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique Role ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Personal data required to register a user.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleSchema"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full role details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "tags": ["Role"],
        "summary": "Delete a role by its role ID.",
        "description": "Delete a role and it's user mappings by the role ID.",
        "operationId": "DeleteRoleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique Role ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full role details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "responses": {
...
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Failed validation on the server",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/FieldValidationError"
            }
          }
        }
      }
    },
    "schemas": {
      "FieldValidationError": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "propertyPath": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "SuccessMessagePair": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "message": {
            "type": "string",
            "default": "Status of the request/operation"
          }
        }
      },
...
      "RoleSchema": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "rootName": {
            "type": "string"
          },
          "viewName": {
            "type": "string"
          },
          "isUserDeletable": {
            "type": "boolean",
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": []
}